Increased the precision of the progress bar display
Request: None

+++ launcher/src/main/java/io/bdeploy/launcher/cli/branding/SplashGraphics.java
@@ -140,7 +140,7 @@ final class SplashGraphics extends JPanel implements LauncherSplashDisplay {
             if (progressCurrent >= progressMax) {
                 fillWidth = progress.width;
             } else {
-                float factor = (float) progressCurrent / (float) progressMax;
+                double factor = (double) progressCurrent / (double) progressMax;
                 fillWidth = (int) (progress.width * factor);
             }
